# x-main/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type TemplateAppJson = { [key: string]: unknown };

export interface Template {
  id: number;
  name: string;
  slug: string;
  description: string;
  githubRepo: string;
  /** @nullable */
  thumbnail?: string | null;
  category: string;
  appJson: TemplateAppJson;
  isFree: boolean;
  isFeatured: boolean;
  showDeployCount: boolean;
  price: number;
  currency: string;
  /** @nullable */
  pairSiteUrl: string | null;
  createdAt: string;
  deployCount?: number;
}

export type TemplateInputAppJson = { [key: string]: unknown };

export interface TemplateInput {
  name: string;
  slug?: string;
  description: string;
  githubRepo: string;
  /** @nullable */
  thumbnail?: string | null;
  category: string;
  appJson: TemplateInputAppJson;
  isFree?: boolean;
  isFeatured?: boolean;
  showDeployCount?: boolean;
  price?: number;
  currency?: string;
  /** @nullable */
  pairSiteUrl?: string | null;
}

export type TemplateUpdateAppJson = { [key: string]: unknown };

export interface TemplateUpdate {
  name?: string;
  description?: string;
  githubRepo?: string;
  /** @nullable */
  thumbnail?: string | null;
  category?: string;
  appJson?: TemplateUpdateAppJson;
  isFree?: boolean;
  isFeatured?: boolean;
  showDeployCount?: boolean;
  price?: number;
  currency?: string;
  /** @nullable */
  pairSiteUrl?: string | null;
}

export type DeploymentStatus = typeof DeploymentStatus[keyof typeof DeploymentStatus];


export const DeploymentStatus = {
  queued: 'queued',
  building: 'building',
  online: 'online',
  offline: 'offline',
  error: 'error',
  failed: 'failed',
  suspended: 'suspended',
} as const;

export type DeploymentEnvVars = {[key: string]: string};

export interface Deployment {
  id: number;
  userId: number;
  templateId: number;
  templateName: string;
  botName: string;
  /** @nullable */
  herokuAppId?: string | null;
  status: DeploymentStatus;
  /** @nullable */
  failedAt?: string | null;
  /** @nullable */
  failureReason?: string | null;
  /** @nullable */
  herokuDeletedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
  /** @nullable */
  templateThumbnail?: string | null;
  envVars?: DeploymentEnvVars;
  /** @nullable */
  databaseUrl?: string | null;
}

export type DeploymentInputEnvVars = {[key: string]: string};

export interface DeploymentInput {
  templateId: number;
  botName: string;
  envVars: DeploymentInputEnvVars;
}

export interface DeploymentLogs {
  lines: string[];
}

export type EnvVarsUpdateEnvVars = {[key: string]: string};

export interface EnvVarsUpdate {
  envVars: EnvVarsUpdateEnvVars;
}

export interface DashboardSummary {
  totalDeployments: number;
  onlineCount: number;
  offlineCount: number;
  errorCount: number;
  recentDeployments: Deployment[];
}

export interface AdminStats {
  totalUsers: number;
  totalDeployments: number;
  onlineDeployments: number;
  totalTemplates: number;
}

export interface FetchAppJsonInput {
  repoUrl: string;
}

export type AppJsonResultEnv = {[key: string]: { [key: string]: unknown }};

export type AppJsonResultRaw = { [key: string]: unknown };

export interface AppJsonResult {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logo?: string | null;
  keywords?: string[];
  env?: AppJsonResultEnv;
  raw: AppJsonResultRaw;
}

export interface PaymentInitiateInput {
  email: string;
  amount: number;
  planType: string;
  currency?: string;
  callbackUrl?: string;
}

export interface PaymentInitiateResponse {
  authorizationUrl: string;
  reference: string;
  accessCode: string;
}

export interface StkPushInput {
  phone: string;
  amount: number;
  email: string;
  planType?: string;
}

export interface MpesaPushResult {
  status: string;
  reference: string;
  message?: string;
}

export interface PaymentVerifyResponse {
  status: string;
  reference: string;
  amount?: number;
  email?: string;
  /** @nullable */
  paidAt?: string | null;
}

export type ListTemplatesParams = {
category?: string;
search?: string;
};