# x-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JuneXDeployment API — Discord bot hosting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}/role:
    patch:
      operationId: updateAdminUserRole
      tags: [admin]
      summary: Update a user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [user, admin]
      responses:
        "200":
          description: Updated account role
          content:
            application/json:
              schema:
                type: object
                required: [id, username, email, role]
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                    enum: [user, admin]
        "400":
          description: Invalid role or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: At least one admin must remain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/fetch-app-json:
    post:
      operationId: fetchAppJson
      tags: [admin]
      summary: Fetch and parse app.json from a GitHub repo URL (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchAppJsonInput"
      responses:
        "200":
          description: Parsed app.json data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppJsonResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/deployments/{id}/suspend:
    post:
      operationId: suspendDeployment
      tags: [admin]
      summary: Suspend a deployment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Deployment:
      type: object
      required: [id, userId, templateId, templateName, botName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: integer
        templateName:
          type: string
        botName:
          type: string
        herokuAppId:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, building, online, offline, error, failed, suspended]
        failedAt:
          type: [string, "null"]
        failureReason:
          type: [string, "null"]
        herokuDeletedAt:
          type: [string, "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        templateThumbnail:
          type: ["string", "null"]
        envVars:
          type: object
          additionalProperties:
            type: string
        databaseUrl:
          type: [string, "null"]
...
    FetchAppJsonInput:
      type: object
      required: [repoUrl]
      properties:
        repoUrl:
          type: string
    AppJsonResult:
      type: object
      required: [name, raw]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: object
        raw:
          type: object
...